Editing RedTypes lib

+++ examples/RedTypesEx.cpp
@@ -5,7 +5,7 @@
 
 int main() {
     Red::int_t a = 4;
-    Red::uint64_t b = 18446744073709551615;
+    Red::uint32_t b = 18446744073709551615;
 
     std::cout << "a: " << a << std::endl;
     std::cout << "b: " << b << std::endl;
